% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_fit_greedy.R
\name{ebpmf_fit_greedy}
\alias{ebpmf_fit_greedy}
\title{ebpmf_fit_greedy}
\usage{
ebpmf_fit_greedy(
  X,
  Kmax = 100,
  ebnm_fn = NULL,
  tol = 0.001,
  maxiter.mu = 10,
  est.tau.dim = 0,
  backfit_first_factor = FALSE
)
}
\arguments{
\item{backfit_first_factor}{whether to backfit the first factor after every iteration
I thought it might be useful if ebnm_fn is non-negative as otherwise the greedy fit
may stop prematurely. but it turns out the just want to keep adding factors in this case so not recommended. Could maybe be removed
and not recommended for now.}
}
\description{
ebpmf_fit_greedy
}
\details{
adds one factor at a time using flash or svd (if ebnm_fn = NULL)
The first factor is fit by svd, so unconstrained.
}
