% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_init_udv.R
\name{ebpmf_init_udv}
\alias{ebpmf_init_udv}
\title{ebpmf_init_udv}
\usage{
ebpmf_init_udv(fit, nu)
}
\arguments{
\item{fit}{an existing ebpmf fit}

\item{nu}{number of singular vectors (columns of u,v) to use}
}
\description{
ebpmf_init_udv
}
\details{
Initializes the u element (only) of the udv list by randomly drawing any needed additional columns for u as white noise
(If udv is already specified, and nu is less than the number of columns of u, then the initialization removes the excess columns)
}
