% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_obj.R
\name{ebpmf_obj}
\alias{ebpmf_obj}
\title{ebpmf_obj}
\usage{
ebpmf_obj(fit)
}
\arguments{
\item{fit}{a ebpmf fit}
}
\description{
ebpmf_obj
}
\details{
returns the objective function for the current fit. Note that
the objective function for SVD-based and flash-based fits may not be comparable.
}
