% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpmf_svd.R
\name{ebpmf_svd}
\alias{ebpmf_svd}
\title{Simple SVD version of EBPMF}
\usage{
ebpmf_svd(
  X,
  nu,
  maxiter = 100,
  tol = 0.001,
  method = c("naive", "als"),
  maxiter.mu = 10,
  est.tau.dim = 0
)
}
\arguments{
\item{X}{a matrix of counts (or a previous ebpmf fit to such a matrix, used as an initialization)}

\item{nu}{the number of singular vectors to compute}

\item{maxiter}{maximum number of iterations}

\item{tol}{convergence tolerance}

\item{method}{string indicating method to use for update, either "als" (alternating least squares) or "naive" (which directly calls svd)}
}
\description{
Simple SVD version of EBPMF
}
\details{
Runs a version of EBPMF where the prior families for L and F are point masses
The EBMF step is equivalent to performing an SVD
}
\examples{
set.seed(1)
n= 10
p = 20
K = 4
LL = matrix(rnorm(n*K),nrow=n)
FF = matrix(rnorm(p*K),nrow=p)
mu = matrix(LL \%*\% t(FF),ncol=p, nrow=n)
X= matrix(rpois(n*p,exp(mu)), ncol=p, nrow=n)
fit = ebpmf_svd(X,nu=5)
fit$tau # should be big
plot(fit$obj) # should show converged
}
